+++ services/Tests/Youverse.Hosting.Tests/YouAuthApi/Drive/DriveQueryTests.cs
@@ -106,7 +106,7 @@ namespace Youverse.Hosting.Tests.DriveApi.YouAuth
 
                 var resultOptions = new QueryModifiedResultOptions()
                 {
-                    MaxDate = (UInt64)DateTimeOffset.UtcNow.AddHours(-1).ToUnixTimeMilliseconds(),
+                    MaxDate = DateTimeOffset.UtcNow.AddHours(-1).ToUnixTimeMilliseconds(),
                     MaxRecords = 10,
                     IncludeJsonContent = false
                 };
